pyproject.toml
[tool.poetry]
name = "babelon"
version = "0.2.1"
description = "babelon"
authors = ["Vinicius de Souza <souzadevinicius@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "babelon", from = "src"}]

[tool.poetry.dependencies]
python = "^3.9"
click = "*"
importlib-metadata = "^4.8.0"
rdflib = "^6.3.2"
jsonasobj2 = "^1.0.4"
xmltodict = "^0.13.0"
pandas = "^2.0.2"
oaklib = "^0.5.9"
tabulate = "^0.9.0"
llm = "^0.13.1"
python-dotenv = "^1.0.1"
curies = "^0.7.7"
sssom = "^0.4.4"
linkml-runtime = "^1.7.1"
linkml = "^1.7.4"

[tool.poetry.group.docs.dependencies]
mkdocs = "^1.4.2"
pymdown-extensions = "^9.9"
mkdocs-material = "^8.5.11"
mkdocstrings = "^0.19.1"
mkdocstrings-python = "^0.8.2"
mkdocs-include-dir-to-nav = "^1.2.0"
mkdocs-click = "^0.8.0"

[tool.poetry.dev-dependencies]
pytest = ">=7.1.2"
tox = ">=3.25.1"
sphinx = {version = ">=6.1.3", extras = ["docs"]}
sphinx-rtd-theme = {version = ">=1.0.0", extras = ["docs"]}
sphinx-autodoc-typehints = {version = ">=1.2.0", extras = ["docs"]}
sphinx-click = {version = ">=4.3.0", extras = ["docs"]}
myst-parser = {version = ">=0.18.1", extras = ["docs"]}


[tool.poetry.scripts]
babelon = "babelon.cli:babelon"

[tool.poetry.extras]
docs = [
    "sphinx",
    "sphinx-rtd-theme",
    "sphinx-autodoc-typehints",
    "sphinx-click",
    "myst-parser"
    ]


[tool.black]
line-length = 100
target-version = ["py38", "py39", "py310"]

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 100
include_trailing_comma = true
reverse_relative = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
